refactor options.js and voxulus.js
