+++ manifest.json
@@ -23,7 +23,6 @@
     },
     "options_page": "/html/options.html",
     "host_permissions": [
-      "*://*\/*",
-      "*/html/options.html"
+      "*://*\/*"
     ]
   }
