+++ scripts/utilities.js
@@ -1,0 +1,16 @@
+async function getCurrentTab() {
+  let queryOptions = { active: true, currentWindow: true };
+  let [tab] = await chrome.tabs.query(queryOptions);
+  return tab;
+}
+
+async function executeScript(func, args) {
+  let tab = await getCurrentTab();
+  chrome.scripting.executeScript({
+    target: { tabId: tab.id },
+    func: func,
+    args: args,
+  });
+}
+
+export { getCurrentTab, executeScript };
